mainViewHandlers are done ish time to test on dev.
