+++ frontend/package.json
@@ -2,34 +2,23 @@
 	"name": "gamechanger-web",
 	"version": "0.1.0",
 	"private": true,
-	"pnpm": {
-		"peerDependencyRules": {
-			"allowedVersions": {
-				"react": "17",
-				"react-dom": "17"
-			}
-		},
-		"overrides": {
-			"bezier-js": "4.0.3",
-			"immer": "9.0.6",
-			"node-fetch": "2.6.1",
-			"node-forge": "1.0.0",
-			"browserslist": "4.17.3",
-			"set-value": "4.0.1",
-			"nth-check": "2.0.1",
-			"ansi-regex": "5.0.1",
-			"glob-parent": "5.1.2",
-			"tar": "4.4.18",
-			"ua-parser-js": "0.7.30",
-			"@types/node": "12.20.13",
-			"@testing-library/react": "12.0.0",
-			"webpack": "4.44.2",
-			"cookie": "0.5.0"
-		}
+	"resolutions": {
+		"bezier-js": "4.0.3",
+		"immer": "9.0.6",
+		"node-fetch": "2.6.1",
+		"node-forge": "^1.2.1",
+		"browserslist": "4.17.3",
+		"set-value": "4.0.1",
+		"nth-check": "2.0.1",
+		"ansi-regex": "5.0.1",
+		"glob-parent": "5.1.2",
+		"tar": "4.4.18",
+		"ua-parser-js": "0.7.30",
+		"types/node": "12.20.13",
+		"testing-library/react": "12.0.0"
 	},
 	"dependencies": {
-		"@babel/core": "7.17.10",
-		"@datapunt/matomo-tracker-react": "0.5.1",
+		"@datapunt/matomo-tracker-react": "0.3.1",
 		"@date-io/date-fns": "1.3.13",
 		"@dod-advana/advana-jira-feedback": "0.4.15",
 		"@dod-advana/advana-platform-ui": "1.4.4",
@@ -46,8 +35,6 @@
 		"@mui/icons-material": "5.3.0",
 		"@mui/material": "5.3.0",
 		"@mui/utils": "5.3.0",
-		"@popperjs/core": "^2.11.5",
-		"@testing-library/dom": "^8.13.0",
 		"anchorme": "2.1.2",
 		"axios": "0.24.0",
 		"bezier-js": "4.0.3",
@@ -63,17 +50,16 @@
 		"history": "4.10.1",
 		"html-react-parser": "1.2.4",
 		"html2canvas": "1.4.0",
-		"jquery": "3.6.0",
 		"js-file-download": "0.4.12",
 		"jws": "3.2.2",
 		"lodash": "4.17.21",
 		"material-ui": "0.20.2",
 		"moment": "2.20.1",
 		"node-fetch": "2.6.1",
-		"node-forge": "1.0.0",
+		"node-forge": "1.2.1",
 		"node-sass": "6.0.1",
+		"npm-force-resolutions": "0.0.10",
 		"piwik-react-router": "0.12.1",
-		"popper.js": "^1.16.1",
 		"prop-types": "15.8.1",
 		"react": "17.0.2",
 		"react-bottom-scroll-listener": "5.0.0",
@@ -81,9 +67,8 @@
 		"react-dom": "17.0.2",
 		"react-error-boundary": "3.1.4",
 		"react-force-graph": "1.41.7",
-		"react-is": "^18.1.0",
 		"react-js-pagination": "3.0.3",
-		"react-loadable": "^5.5.0",
+		"react-loadable": "5.5.0",
 		"react-loadable-visibility": "3.0.2",
 		"react-modal": "3.14.4",
 		"react-motion": "0.5.2",
@@ -97,13 +82,10 @@
 		"sanitize-html": "2.6.1",
 		"spark-md5": "3.0.2",
 		"styled-components": "4.4.1",
-		"three": "0.140.0",
 		"typeface-montserrat": "0.0.54",
 		"typeface-noto-sans": "0.0.54",
 		"underscore": "1.13.2",
 		"uuid": "3.4.0",
-		"webpack": "4.44.2",
-		"worker-loader": "^3.0.8",
 		"xlsx": "0.17.0"
 	},
 	"devDependencies": {
@@ -120,15 +102,18 @@
 		"eslint-plugin-react": "^7.24.0",
 		"prettier": "^2.6.1",
 		"sonarqube-scanner": "^2.5.0",
+		"source-map-explorer": "^2.5.2",
 		"typescript": "^4.4.3"
 	},
 	"scripts": {
 		"109": "REACT_APP_GLUU_SSO=disabled REACT_APP_NODE_ENV=test109 yarn start",
+		"analyze": "source-map-explorer 'build/static/js/*.js'",
+		"preinstall": "([ ! -f package-lock.json ] && npm install --package-lock-only --ignore-scripts --no-audit); npx npm-force-resolutions",
 		"start": "react-scripts start",
 		"dev": "REACT_APP_GLUU_SSO=disabled REACT_APP_NODE_ENV=development PORT=8080 yarn start",
 		"devDecoupled": "REACT_APP_GLUU_SSO=disabled NODE_ENV=development REACT_APP_NODE_ENV=development REACT_APP_GC_DECOUPLED=1 yarn start",
 		"awsDocker": "REACT_APP_GLUU_SSO=disabled REACT_APP_NODE_ENV=aws PORT=8080 react-scripts --max_old_space_size=8096 start",
-		"build": "DISABLE_ESLINT_PLUGIN=true react-scripts --max_old_space_size=16000 build && compress-cra",
+		"build": "DISABLE_ESLINT_PLUGIN=true react-scripts --max_old_space_size=16000 build && npx compress-cra",
 		"test": "jest --watch",
 		"jenkins-test": "jest",
 		"test-coverage": "jest --coverage --coverageDirectory=coverage/jest",
