+++ frontend/src/components/modules/globalSearch/globalSearchMainViewUtilityHandler.js
@@ -1,7 +1,11 @@
 import { setState } from '../../../utils/sharedFunctions';
 
 const GlobalSearchMainViewUtilityHandler = {
-	handleCategoryTabChange: (props) => {},
+	handleCategoryTabChange: (props) => {
+		const { tabName, dispatch } = props;
+
+		setState(dispatch, { activeCategoryTab: tabName, resultsPage: 1 });
+	},
 };
 
 export default GlobalSearchMainViewUtilityHandler;
