--- frontend/src/components/modules/jbook/jbookMainViewUtilityHandler.js
@@ -1,7 +1,0 @@
-import { setState } from '../../../utils/sharedFunctions';
-
-const JBookMainViewUtilityHandler = {
-	handleCategoryTabChange: (props) => {},
-};
-
-export default JBookMainViewUtilityHandler;
