--- frontend/src/components/modules/policy/policyMainViewUtilityHandler.js
@@ -1,7 +1,0 @@
-import { setState } from '../../../utils/sharedFunctions';
-
-const PolicyMainViewUtilityHandler = {
-	handleCategoryTabChange: (props) => {},
-};
-
-export default PolicyMainViewUtilityHandler;
